move core logic to util functions
